server: support escaping separator in search terms.
For non-regex search terms, the separator is hardcoded to a comma
character (which is fine), but there has been no way to escape it until
now. Since commas are legal characters in references (and probably other
fields), having a way to search for them can actually be useful.

While one could argue that regular expressions can be used to achieve
that, typical users have no idea how that works and often get them
wrong, so providing an easy alternative sounds beneficial.